requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
alabaster=0.7.12=py_0
argh=0.26.2=py37_0
asn1crypto=1.3.0=py37_0
astroid=2.3.3=py37_0
atomicwrites=1.3.0=py_0
attrs=19.3.0=py_0
autopep8=1.4.4=py_0
babel=2.8.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.4=py_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h2e261b9_0
chardet=3.0.4=py37_1003
cloudpickle=1.3.0=py_0
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
dbus=1.13.12=h746ee38_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20181111=py_0
docutils=0.16=py37_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
flake8=3.7.9=py37_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=py37_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
idna=2.9=py_1
imagesize=1.2.0=py_0
importlib_metadata=1.5.0=py37_0
intel-openmp=2020.0=166
intervaltree=3.0.2=py_0
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
isort=4.3.21=py37_0
jedi=0.15.2=py37_0
jeepney=0.4.3=py_0
jinja2=2.11.1=py_0
jpeg=9b=h024ee3a_2
jsonschema=2.6.0=py37_0
jupyter_client=6.1.2=py_0
jupyter_core=4.6.3=py37_0
keyring=21.1.1=py37_2
kiwisolver=1.1.0=py37he6710b0_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libspatialindex=1.9.3=he6710b0_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h7b6447c_0
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
nbconvert=5.6.1=py37_0
nbformat=5.0.4=py_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py37hfd86e86_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
numpydoc=0.9.2=py_0
olefile=0.46=py37_0
openssl=1.1.1f=h7b6447c_0
packaging=20.3=py_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.2=py_0
pathtools=0.1.2=py_1
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.1.0=py37h34e0f95_0
pip=20.0.2=py37_1
pluggy=0.13.1=py37_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
psutil=5.7.0=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
pycodestyle=2.5.0=py37_0
pycparser=2.20=py_0
pydocstyle=4.0.1=py_0
pyflakes=2.1.1=py37_0
pygments=2.6.1=py_0
pylint=2.4.4=py37_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_0
python=3.7.7=hcf32534_0_cpython
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.3.4=py_0
python-language-server=0.31.9=py37_0
pytorch=1.3.1=cuda100py37h53c1284_0
pytz=2019.3=py_0
pyxdg=0.26=py_0
pyyaml=5.3.1=py37h7b6447c_0
pyzmq=18.1.1=py37he6710b0_0
qdarkstyle=2.8=py_0
qt=5.9.7=h5867ecd_1
qtawesome=0.7.0=py_0
qtconsole=4.7.2=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
requests=2.23.0=py37_0
rope=0.16.0=py_0
rtree=0.9.3=py37_0
scipy=1.4.1=py37h0b6359f_0
secretstorage=3.1.2=py37_0
setuptools=46.1.3=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.1.0=py37_0
sphinx=2.4.4=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.2=py37_0
spyder-kernels=1.9.0=py37_0
sqlite=3.31.1=h7b6447c_0
testpath=0.4.4=py_0
tk=8.6.8=hbc83047_0
torchvision=0.4.2=cuda100py37hecfc37a_0
tornado=6.0.4=py37h7b6447c_1
traitlets=4.3.3=py37_0
ujson=1.35=py37h14c3975_0
urllib3=1.25.8=py37_0
watchdog=0.10.2=py37_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py37_1
wheel=0.34.2=py37_0
wrapt=1.12.1=py37h7b6447c_1
wurlitzer=2.0.0=py37_0
xz=5.2.5=h7b6447c_0
yaml=0.1.7=had09818_2
yapf=0.28.0=py_0
zeromq=4.3.1=he6710b0_3
zipp=2.2.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
